Make testClearSelectionInRecentsResetsActions more robust
If the device returns results that exceed the limits of the recents max,
then sometimes the file isn't returned immediately. Given the fact the
file exists or not is not required, just the selected state, let's
simply select the first file in the list of Recents. Regardless if the
file is visible or not (will fail if NO file is visible still).

Bug: 412776499
Test: atest com.android.documentsui.FilesActivityUiTest
Flag: com.android.documentsui.flags.use_material3
Change-Id: I1a6aa1ff06d81a606ff95cc5dbe073aa99a4779d