Fix up RootsFragment test when use_material3 flag enabled
A previous change migrated the usage of FEATURE_PC to an RRO property to
conditionally show/hide the media roots. Unfortunately it missed some
tests which failed because they took a new code path to get the
resources than the mocked one. This updates the test to just check the
context for the RRO property and assert the order based on the device
that presents itself.

Bug: 412776499
Test: atest com.android.documentsui.sidebar.RootsFragmentTest
Flag: com.android.documentsui.flags.use_material3
Change-Id: I4bacff073435d04ba46290217af08f90b3c1e0f0