Don't expect PHOTOPICKER to appear on devices without it
The TrampolineActivity currently expects the ACTION_PICK_IMAGES to be
available on the device otherwise it opens up DocumentsUI. The way
mainline works in tests is it flashes the base image then the module on
top, so that means for DocsUI tests it sometimes doesn't have
Photopicker available in the environment. Update the test to ensure it
tests the appropriate combination and update the strings so this
information is readily available when debugging the tests.

Bug: 407678392
Test: atest com.android.documentsui.TrampolineActivityTest
Flag: com.android.documentsui.flags.redirect_get_content_ro
Change-Id: I5dd83c56702d9c6d1d813644d4ae0264180edf80