[DocsUI, Search]: Fix failing DocumentsUIUnitTests
The test reports that "No virtual method toSeconds()" exists on the
Duration class or its superclasses. It was implemented in JDK Level 9,
but I guess it did not make it into Android code(??). This cl uses the
default toString() implementation of Duration to avoid any problems. It
is a bit less readable, but more reliable.

Bug: 421965206
Test: m DocumentsUIGoogle
Flag: com.android.documentsui.flags.use_search_v2_read_only
Change-Id: Ie43b016a452591f7afe67d297dd1bbfcdfce78dd