Fix TrampolineActivityTest on Android T
On Android T base image, MediaProvider doesn't provide an enabled
GET_CONTENT handler. This means the tests currently fail as we first
check if the component is enabled or not. Update the tests to align with
the expectation that we don't forward this to Photopicker.

Largely this logic is in TrampolineActivity but we should use the same
logic in the tests to verify the device is valid. This extracts that
logic into a utility class that can be imported by both the tests and
the Activity.

Bug: 421894788
Test: atest com.android.documentsui.TrampolineActivityTest
Flag: com.android.documentsui.flags.redirect_get_content_ro
Change-Id: Ifbe0b5f96137baf8961752e7076480eeff963048