Synchronize access to ResolvedResourcesJob.mResolvedDocs
The mResolvedDocs member is getting accessed in getProgressMessage(),
which is called from the progress reporting thread, a different thread
than the main thread running setUp(), start() and finish().

Since mResolvedDocs is modified in setUp(), proper synchronization is
needed when accessing it in getProgressMessage().

Putting in place this synchronization also allows us to simplify
UnpackJob.

Bug: 422899453
Flag: EXEMPT fix existing synchronization issues
Test: atest DocumentsUIGoogleTests:com.android.documentsui.services
Change-Id: I361f269c9fd0e06507d0ebf8f97761cd0aa9a298