[DocsUI, Search]: Fix private space handling.
Removes FolderInfo, and replaces it with RootInfo. This is due to the
fact that we only do searches from the roots of providers. Rewrites the
logic of the base loader to use the user supplied for each root, rather
than a collected list of users. Adjusts test.

Bug: 417870902
Bug: 423757780
Test: m DocumentsUIGoogle
Flag: com.android.documentsui.flags.use_search_v2_read_only
Change-Id: Iac464f5f0f7c234d801c1438f3661c7ead5827ee