[DocsUI]: Fix testClickRecent failing test.
The test checks if the menu item with ID option_menu_search is visible.
However, in Material3 this is never the case. The menu item is replaced
with docked search toolbar. This CL splits the tests into not-M3 and M3
and adjust them appropriately.

Bug: 426782669
Test: atest DocumentsUIUnitTests
Flag: com.android.documentsui.flags.use_material3
Change-Id: Id8b828b9845297fe43a1a4c8ef39ba69e82e9cdc