[DocsUI]: Fix testFilesList_LiveUpdate test.
The live update test may fail if a small device uses grid display. We
try to minimize the chance of the new file being invisible by explicitly
switching to list mode. However, in addition the assertDocumentsPresent
is renamed to assertDocumentsVisible to make it explicit what it is
testing. A new method, asssertDocumentsPresent is added. Both just call
assertDocumentExistWithScroll that takes scrolling as an additional
parameter.

Bug: 424299272
Test: atest DocumentsUIUnitTests
Flag: com.android.documentsui.flags.use_material3
Change-Id: I9c9352708d27fca06bb270c5e3edeeb09c63da7b