[DocsUI] Fix refresh logic when open in new window
Previously we assign the stack from the intent extra to the existing
stack in the state before calling onRootPicked(), but inside the
onRootPicked() method it will compare the root passed in with the
existing stack first before refreshing the file list, it's always
true so the refreshing is always skipped. This CL removes the
stack.reset() before calling onRootPicked() and relies on the
stack.changeRoot() inside to cover the same functionality.

Bug: 417577791
Test: atest DocumentsUIGoogleTests:com.android.documentsui.SidebarUiTest#testOpenInNewWindow_preservesFiles
Flag: EXEMPT bug fix
Change-Id: I3f562d82b2ceb844bb839daeddba84d403e6d1a2