Remove "open with" on desktop when the file less than 2 opening apps
"open with" opens ResolverActivity which currently automatically starts
the only app when there is a single option to choose from.

There is a plan for 26Q1 to fix this in ResolverActivity but for now we
have to hide "open with" when it would behave unexpectedly.

Bug: 420521558
Test: atest DocumentsUIGoogleTests
Flag: com.android.documentsui.flags.desktop_file_handling_ro

Change-Id: Iab3c5c391c974c84c79c2f1e67f8d313638045a7