[DocsUI] Add SkipScreenRecording annotation for tests
Certain tests, such as those in FileCopyUiTest, do not play nicely
with the test harness screen recording collector. Add annotation
so we can configure TradeFed to not use the recorder for these
tests.

Note that this CL alone is a noop, it requires the other CL in
this topic to take effect.

Flag: EXEMPT bugfix
Bug: 428081621
Test: atest DocumentsUIGoogleTests:com.android.documentsui.FileCopyUiTest
Change-Id: I3e022455a695ff0a92a264c381122101da0e85c8