Add idling resource to JobPanelUiTest
These tests send job progress updates as broadcasts and the UI updates
when it receives one. Thus there is a race if the UI is checked right
after a progress update is broadcasted. So, mark the UI as busy after
sending a job progress update until the broadcast is detected.

Bug: 445594958
Test: atest 'DocumentsUIGoogleTests:com.android.documentsui.JobPanelUiTest'
Flag: com.android.documentsui.flags.visual_signals_ro
Change-Id: I0f903f3834db6ade2c6d35514b5703df295d9aaa