[DocsUI, Search]: Fix testSearchV2LastModifiedDropdownVisibility
The test fails to find the TEST_ROOT_0 on small screens due to it being
below the bottom of the screen. Even though the oepnRoot scrolls, it
does not seem to interact with the directory. Thus switching it to
Downloads. It really does not matter what root we pick, as long as it is
not Recents.

Bug: 428057041
Test: atest DocumentsUIGoogleTests
Flag: com.android.documentsui.flags.use_search_v2_read_only
Change-Id: I479213b8f8e2f836c64702dbe7a6cd437d65ac4a