[DocsUI] Improve stability of SidebarBot
Various tests fail "downstream" from trying to open the navigation tree
drawer but actually failing to (because a click on the hamburger button
doesn't work).

Ensure UiAutomator thinks it can find the hamburger button before
clicking it, fail hard if it doesn't appear, and wait for the root list
to appear after clicking the button.

Add a helper to have Espresso wait for certain UI elements to appear
and use it in combination with a more restrictive view matcher to
ensure that the context menu is actually open before trying to click
in it.

Bug: 398545748
Flag: EXEMPT bugfix
Test: atest DocumentsUIGoogleTests com.android.documentsui.SidebarUiTest#testPasteIntoFolderOnRoot
Change-Id: I79fcaea472cb1c80c510c6e05dab245ab7cef13a