[DocsUI a11y] Expand the collapsing layout when it's focused
When CollapsingToolbarLayout is collapsed, focusing its children
which are scrolled above the screen won't make it expand, which
makes the focused element not visible, causing an a11y issue.

This CL will expand the layout when:
* any child views inside CollapsingToolbarLayout is focused, and
* the layout is in collapsed state.

Check the attached bug for video demo.

Bug: 431953449
Test: atest DocumentsUIGoogleTests:com.android.documentsui.NavigationViewManagerTest
Flag: com.android.documentsui.flags.use_material3
Change-Id: Id21b7fd12a69b13e90bfacf843dcea4f8592d6d2