[DocsUI Home screen] Add ShortcutItem and BaseSidebarEntryItem.
See go/homescreen-files-docsui-dd.

ShortcutItem is introduced to allow the shortcuts to appear on
the sidebar of DocsUI. This item would look visually similar and
will have shared behaviour to the existing RootItem class, so
we introduce the BaseSidebarEntryItem class to be an overarching
class over ShortcutItem and RootItem. Existing methods in RootItem
that implement the shared behaviours will be moved over to
BaseSidebarEntryItem.

Bug: 438374404
Test: atest DocumentsUIGoogleTests:com.android.documentsui.sidebar.BaseSidebarEntryItemTest
Flag: EXEMPT PURE_REFACTOR
Change-Id: Ib4cf45e9968de903c5a708f9a0e94afbdc6dfab2