[DocsUI Home screen] Refactor GetRootDocumentTask.
The implementation of GetRootDocumentTask can be used to get any
document info as long as the authority, document id and user id
supplied is valid. Refactor this task to take these arguments in
as parameters instead of just a RootInfo so it can be reused in
other places.

Note: TestDocumentsAccess.getRootDocument() is never used so there
will be no test changes for this method.

Bug: 439693594
Test: EXEMPT PURE_REFACTOR
Flag: EXEMPT PURE_REFACTOR
Change-Id: I463efb1c4267e53c703012e06696e2dcfdb6ce73