[DocsUI, Search]: Fix folder loader cancelation.
Exposes cancel signal to classes extending BaseFileLoader. Adds creation
of a cancel signal to FolderLoader so that it can be cancelled once
cancelLoadInBackground is called. Adds a countdown latch to
TestDocumentsProvider so that we know when loading is paused. Adds a
test that checks that cancelling a loader works as intended.

Bug: 435523675
Flag: com.android.documentsui.flags.use_search_v2_read_only
Test: atest DocumentsUIGoogleTests
Change-Id: I10c49cc49c4b74ff37aad17f39e247f7e5c3a2f9