[DocsUI, Search]: Adds test for search view.
Adds a generic assumption testing that checks if the given predicate
returns true. If not it fails the assumption with the given message.
Uses the added assumption code to check for small and large windows.
Given the asumption is true, it checks that the user interface of the
search view matches the expectations.

Bug: 435522285
Bug: 435522307
Flag: com.android.documentsui.flags.use_search_v2_read_only
Test: atest DocumentsUIGoogleTests
Change-Id: I55838a4b41b59bff8629e27cef229a11c0765e93