[DocsUI Home screen] Refactor isValidDestination()
Pass in the list of invalid destinations for the call. The
current implementation can cause issues with asynchronous
calls that will be implemented in future CLs. In detail,
mInvalidDestination can get reset to null before isValidDestination()
is called from an asynchronous task. The asynchronous task would however
require the list of invalid destinations that was present when the task
was first started, not the null list.

Bug: 441194501
Test: EXEMPT PURE_REFACTOR
Flag: EXEMPT PURE_REFACTOR
Change-Id: I8299c718db09045afd5fbbc4f6fea2a379f955f9