[DocsUI Home screen] Update items at the beginning of breadcrumb path
See go/homescreen-files-docsui-dd.

The show() method of the breadcrumb does not allow the items at the
very front of the breadcrumb to be updated when switching to/from
shortcuts. Updates to the beginning of the breadcrumb path have to be
directly called and added in this CL.

Bug: 441194838
Test: atest DocumentsUIGoogleTests:com.android.documentsui.FilesActivityUiTest
Flag: com.android.documentsui.flags.home_screen_files_ro
Flag: com.android.documentsui.flags.use_material3
Change-Id: I4b88760d7bd64f3780af004f967b158155b8c77e