[DocsUI M3] Remove the tab cycle UI test
When the first tab is pressed the focus should go to the side
navigation bar ideally but it can not be guaranteed due to the
existence of the grey overlay on the main coordinator_layout,
which causes the flakiness. Several efforts have been tried to
de-flake it, e.g. focus on the coordinator_layout before pressing
tab, or focus on the navigation root list first before pressing the
tab, unfortunately they don't actually help: for one it's still
flaky after these mitigations, for two, it now skips the first
tab press and only tests the tab cycle in navigation roots (the
test only covers 2 roots) which doesn't cover much.
Given all these reason, remove the test and we will revisit the test
after removing the grey overlay in b/417871278.

Bug: 432508904
Test: EXEMPT remove test
Flag: com.android.documentsui.flags.use_material3
Change-Id: I464ae7fd63d762d55de4bab859e9ea8881445ee4