[DocsUI M3] Fix selection bar related UI test
Previously the toolbar still shows even if the selection bar shows, on
the UI the selection bar completely covers the toolbar. A lot of UI
tests rely on clicking the 3-dot menu from the toolbar even when the
selection bar shows, CL:35418669 hide the toolbar when the selection
bar shows which breaks all these tests.

This CL updates all the relevant clickToolbarOverflowItem() to
clickActionbarOverflowItem() to make sure they are actually operating
on the action bar (selection bar) 3-dot menu, which is consistent with
what end user does on the UI.

Bug: 442574027
Test: atest DocumentsUIGoogleTests:com.android.documentsui.CancelFromNotificationUiTest
Test: atest DocumentsUIGoogleTests:com.android.documentsui.FileCopyUiTest
Test: atest DocumentsUIGoogleTests:com.android.documentsui.FileDeleteUiTest
Test: atest DocumentsUIGoogleTests:com.android.documentsui.FileManagementUiTest
Test: atest DocumentsUIGoogleTests:com.android.documentsui.ArchiveUiTest
Test: atest DocumentsUIGoogleTests:com.android.documentsui.FilesActivityUiTest
Flag: EXEMPT testing code
Change-Id: I6b94b72f64891e896bbe40d0ad370afabd86e2b5