[DocsUI Home screen] Opening shortcut sidebar entry
See go/homescreen-files-docsui-dd.

New implementation when clicking on a shortcut sidebar entry. The new
implementation should open the contents of the shortcut folder upon a click
If the shortcut folder does not exist yet, a task will be started to
create the folder of the shortcut and open to that folder.

Bug: 441193752
Test: atest DocumentsUIGoogleTests:com.android.documentsui.FilesActivityUiTest
Flag: com.android.documentsui.flags.home_screen_files_ro
Change-Id: I6c0f4762e6bf11eb759460538d00ab94e5df255d