[DocsUI, Search]: Fix crash on restore.
Prevents onMove to be executed when the cursor is closed. When the
previous loader is closed, DirectoryResults are closing the cursor
before the new loader has a chance to update the model with the new
DirectoryResult. As a consequence BulkCursorToCursorAdaptor was throwing
an exception. This CL prevents the parent's class method to be involved
on already closed cursor.

Bug: 442970558
Test: manual
Flag: com.android.documentsui.flags.use_search_v2_read_only
Change-Id: I573698aed38d9425172d4e0cacf1fea992e2f915