[DocsUI, Search]: Adds path extractor.
The job of a path extractor is to extract a human readable full path for
a given DocumentInfo. This CL only adds the class and tests. In
subsequent CL, the path returned by the path extractor is going to be
set on the breadcrumb view.

Bug: 416108180
Flag: com.android.documentsui.flags.use_search_v2_read_only
Test: atest DocumentsUIGoogleTests
Change-Id: I313c8131c40222943b872a7e393027a141a3a140