[DocsUI, Search]: Show path of search results.
Adds connection between directory fragment and breadcrumb view. This is
done by listening to selection events in the DirectoryFragment class. If
one element is selected, and the directory fragment has access to
breadcrumb model and the path, it updates the breadcrumb model with the
path extracted by the path extractor. New framework is added for testing
breadcrumb v2 to BreadBot. This is used in SearchViewUiTest to check
that the correct path has been shown to the user on selection events.

Bug: 416108180
Flag: com.android.documentsui.flags.use_search_v2_read_only
Test: atest DocumentsUIGoogleTests
Change-Id: Id5c6d6916bcb0fd9ec2966bb2963c1a03e7ff4af