[DocsUI] Fix NavRailUiTest on devices with smaller size
In the test we are launching the app windows at coordinate (0,0) with
size 700x900dp, because we only run the test if the screen window is
bigger than this. However, in some devices the screen size is higher
than 900, but the bottom of Files app is covered by the navigation bar,
that's why ROOT_1 is not showing there because it's covered. This CL
adds a 100dp buffer when checking the screen height, to make sure the
app window can get the specific size we set and not being covered by
the system bars.

Bug: 445666318
Test: atest DocumentsUIGoogleTests:com.android.documentsui.NavRailUiTest
Flag: com.android.documentsui.flags.use_material3
Change-Id: I41f1ad5c80c39848b7d6e6d9608c0de4ae6234ae