[DocsUI] Make ActionCreateDocumentsUiTest use ActivityTesJunit4
The test was extending from DocumentsUiTestBase before, whose
functionalities are already covered by the standard ActivityTestJunit4.
This CL deletes the DocumentsUiTestBase (it's only used here), and
extends from ActivityTesJunit4 instead. This is also an attempt to
fix the flaky issue of the test because it complained about
"Activity wasn't created yet or already destroyed" before.

Bug: 444280527
Test: atest DocumentsUIGoogleTests:com.android.documentsui.ActionCreateDocumentUiTest
Flag: EXEMPT TEST_ONLY
Change-Id: I1f7cf721c3f7bb3f0c0ccbef0524dc765de37e66