[DocsUI] cancel test notifications w/o progressbar
This is a test only change.

CancelFromNotificationUiTest.java was added in 2017
(https://android-review.git.corp.google.com/c/platform/packages/apps/DocumentsUI/+/545099)
and adds 4 tests: a 2x2 matrix where one axis is "copy or move" and the
other axis is "cancel from the picker dialog's Cancel button or cancel
from the notification after choosing the picker dialog's OK button".

Prior to this commit, the two "from the notification" tests were flaky
(as of September 2025). The TNS (TestNotificationService) doCancel
method calls isStartProgress and so it won't actually execute the
notification's cancel action unless that notification has a progress
bar.

In 2025, on modern hardware, it is possible for the actual copying
action to happen so quickly, and/or there are differences now in who
wins IPC races between:
- the DocumentsUI process (where notifications are created),
- the TNS process (where notifications are processed and their cancel
  action executed, via an intent) and
- the FileOperationService process (where the file copying happens)

so that the TNS process only *sometimes* sees a notification *with a
progress bar*. "Sometimes" might have been "very frequently" in 2017
when the (non-deterministic) test was written but, after all the code
changes since then (in DocumentsUI but also in Android, whether UI
changes, process scheduler or I/O changes or something else) or hardware
changes, it is not always "very frequently" now.

This commit simply removes the TNS requirement that, if we're in
MODE.CANCEL_MODE (instead of the only other mode, MODE.EXECUTION_MODE),
a notification has to have a progress bar for onNotificationPosted to
actually do anything.

There are only three test files use that TNS:
- CancelFromNotificationUiTest.java uses MODE.CANCEL_MODE and
- the other two (FileCopyUiTest.java and FileDeleteUiTest.java) use
  MODE.EXECUTION_MODE.

On my hardware, "atest
DocumentsUIGoogleTests:com.android.documentsui.CancelFromNotificationUiTest"
would reliably fail before this commit and pass afterwards.

CancelFromNotificationUiTest.java's
testCopyDocument_CancelFromNotification will assertFalse that the
destination directory (the ROOT_1_ID directory) will contain the copied
file. It asserts that the copy-the-file job will clean up partial state
if cancelled. Indeed, even after this commit, commenting out the "if
(!success) { deleteDocument(); }" fragment in CopyJob.java's
copyFileHelper method will cause testCopyDocument_CancelFromNotification
to fail on my hardware. In other words, after this commit, the test can
still fail if the CopyJob does not clean up partial state if cancelled.

Bug: 427618116
Bug: 437236527
Flag: EXEMPT TEST_ONLY
Test: atest DocumentsUIGoogleTests:com.android.documentsui.CancelFromNotificationUiTest
Change-Id: I802534c6b5f704622f1534f9175de2b1f9cf24bf