[DocsUI] Fix Espresso item matching and RootsAdapter for shortcuts
When opening a root, the current implementation of EspressoBots
only tries to match the label to RootItem instances when the
UseMaterial3 flag is off. We now enable the label to be matched
to BaseSidebarEntryItem instances to allow shortcuts to be matched.

RootsAdapter also modified to account for and handle shortcuts when
the material3 flag is off.

Bug: 447529706
Bug: 441193752
Test: atest DocumentsUIGoogleTests:com.android.documentsui.FilesActivityUiTest
Flag: com.android.documentsui.flags.home_screen_files_ro
Change-Id: If4f79359506904dd8ddf61a06bf575d853c208c9