[DocsUI] Fix LayoutUtils checks for PickActivity
The tests assume that the existing activity is always a FilesActivity.
However, tests under PickActivity should check for the layout of the
PickActivity as it could differ from the layout of the FilesActivity
under certain screen dimensions. This CL will allow PickActivity tests
to check for the PickActivity layout rather than using the
FilesActivity layout.

Bug: 447066451
Flag: EXEMPT TEST_ONLY
Test: atest DocumentsUIGoogleTests
Change-Id: Ifcb7fae8ba4377bb9ccbc207b92001bed597b21f