[DocsUI] Invalidate decorations after observing no scroll/layout
This is a follow up to ag/35697339 that tried to fix the crash that
occurs when invalidateItemDecorations is called during a scroll or
layout. The fix by simply posting invalidateItemDecorations to the
UI thread was not sufficient.

Create the ItemDecorationInvalidator to explicitly observe the
scroll and layout state and call invalidateItemDecorations when
they are idle.

The only way to observe the recycler view layout state is to use a
ViewTreeObserver.OnGlobalLayoutListener and this is noisy
because it's called whenever any layout on the view tree
changes, not just for recycler view. Further, the check for
idleness upon a layout state change needs to be in a runnable
that's posted to the UI thread to ensure that the layout has
fully finished first.

To minimise spamming the UI thread, only observe the layout state when
scrolling is idle, because a scroll will likely lead to a layout anyway.
Also only post one runnable at a time.

Flag: com.android.documentsui.flags.use_material3
Bug: 415947063
Bug: 415947192
Bug: 448038349
Test: atest com.android.documentsui.SearchViewUiTest#testSearchV2RootNameIsAdjusted
Test: atest com.android.documentsui.dirlist.ItemDecorationInvalidatorTest
Test: atest com.android.documentsui.dirlist.DirectoryFragmentTest
Change-Id: Ieeae9a510fe527df6e0973f1927ff8b84cd4fba2