[DocsUI M3] Fix inconsistent padding for browser/picker
When the app is rendered in fullscreen mode, the UI will be overlapped
with the system bars (e.g. top status bar and bottom navigation bar),
we need to set additional padding to avoid the overlap. Previously in
BaseActivity this was done in a wrong way:
* The padding was set on multiple child views unnecessarily.
* The padding was not being reset when navigation bar is hidden.

This CL uses the system bar insets which caters both status bar and
navigation bar, and also sets the padding on the root container
instead. In addition, when navigation gesture is used, we just use
that transparent area as bottom padding to achieve a immersive
experience.

For PickActivity, when it's rendered in the dialog mode, instead of
setting the padding of the root container, we need to set the margin
because it's the margin area which overlaps with the system bars.

Check the attached bug for screenshots.

Bug: 448496264
Test: m DocumentsUIGoogle && manual inspection
Flag: com.android.documentsui.flags.use_material3
Flag: com.android.documentsui.flags.desktop_ux_phase_2_ro
Change-Id: I94884c48c4b31bf7e5785b129c05d9f3f63276f0