DocsUI: Initiate the summary loader and display in the file list
Change the AbstractActionHandler to start the SummaryLoader and push the
fetched summaries to Model.

The model when receives the fetched summaries it propagates to the
RecyclerView adapter (ModelBackedDocumentsAdapter) which is a listener
to the newly added event ModelSummariesUpdate (in the prior CL).

Bug: 436750999
Flag: com.android.documentsui.flags.use_file_summary
Test: DocumentsUIGoogleTests

Change-Id: Ia272cc662c2e572031eced82dd715cd8fa18cead