Fix incorrect down timing on GestureBot events
On DOWN, the mDownTime variable isn't properly set meaning the MOVE/UP
events have an incorrect down timestamp of 0. This is rejected by the
input verifier as the event timestamp goes backwards. This variable
should also not be reset on UP as it should never decrease.

Bug: 450445268
Test: atest 'DocumentsUIGoogleTests:com.android.documentsui.BandSelectionUiTest'
Flag: EXEMPT TEST_ONLY
Change-Id: Ie1dc341dd361920d8111543e6903f553ff3ccb86