[DocsUI, Search]: Breadcrumb driven directory change
This patch enables users to navigate to a directory by clicking on
an item in the search results breadcrumb.

In DirectoryFragment, when a user selects a search result, a
ClickConsumer is now attached to the BreadcrumbController. When a
breadcrumb item is clicked, this listener resets the main activity's
DocumentStack to the path of the selected file and calls a new
onDirectoryChange method. This new method triggers navigation by
calling the new onNavigateToStackItem method in NavigationViewManager.

To support this, PathExtractor is refactored to return a full
DocumentStack object instead of a string array, and it now requires
a ProvidersAccess dependency. Consequently, BaseActivity adds a
getProvidersAccess method to supply this dependency. Finally, ownership
of the BreadcrumbController is moved from NavigationViewManager to
BaseActivity to make it centrally accessible.

Bug: 447678204
Flag: com.android.documentsui.flags.use_search_v2_read_only
Test: atest DocumentsUIGoogleTests
Change-Id: I968d799ed56ec5eb0177568778f73db54708c951