[DocsUI M3] Fix the wrong selection bar color in picker
When picker was rendered as a dialog, somehow the selection bar shows
a grey-ish background color initially, the color will turn back to
normal when resizing to other layouts (a relayout happens). Not sure
why but after removing the explicit background color attribute on the
selection bar, the issue is gone. Removing the explicit background color
won't change anything, the selection bar will just pick the container
background color, which is what we want (similar to the toolbar).

Bug: 448496544
Test: m DocumentsUIGoogle && manual inspection
Flag: com.android.documentsui.flags.use_material3
Change-Id: Iac14f77638f1945486c9672a3f0abcb020b64af8