Hide trash-related items by default
The "Move to trash" and "Restore from trash" menu items were incorrectly visible by default in the action and context menus. This was because the default visibility of menu items is true.
This change fixes the issue by setting `android:visible="false"` for these menu items in the XML layout. This ensures that the items are not visible unless explicitly enabled.

Bug: 451418059
Bug: 451495501
Bug: 451413839
Flag: com.android.documentsui.flags.enable_trash_flow_ro
Test: atest DocumentsUIGoogleTests:com.android.documentsui.files.MenuManagerTest
Change-Id: I87f285718452debe0884bcbc8b1d2dc96cec9dbc