[DocsUI, Search]: Quick fix for broken options search.
In ag/36248687 changing search query and posting notifications was
factored out to the setCurrentSearch. If this method detected a change
in the search query, it would update the SearchManagerListener. However,
that refactoring did not take into account the case when options have
changed. In such a case, the query does not change, but the search is
still triggered. This change fixes the issue by adding a return that
indicates if setCurrentSearch posted a notification. In search task if
search was started and notification was not sent we generate extra
notification.

Bug: 450381836
Flag: com.android.documentsui.flags.use_search_v2_read_only
Test: atest DocumentsUIGoogleTests
Change-Id: I8ab2542f864aba583df30b7e02dac46667fa8bf4