[DocsUI M3] only add bottom padding for the right section
Remove the main_container ids for fixed_layout and nav_rail_layout
because there's a main_container in drawer_layout with a different type
which causes a crash. Instead use container_save (which is located at
the bottom of the right section) to set bottom padding, this container
always shows in the layout regardless of in picker/saver mode or not,
so it's reliable to use it to set bottom padding.

This CL also changes the bottom padding logic: we don't need bottom
padding for the navigation tree area in any cases (including window
mode and fullscreen mode), that's why now root container always has 0
bottom padding.

Check the attached bug for the screenshots.

Bug: 448496264
Bug: 453477875
Test: m DocumentsUIGoogle && manual inspect
Flag: com.android.documentsui.flags.use_material3
Flag: com.android.documentsui.flags.desktop_ux_phase_2_ro
Change-Id: I0eb27d10a938aeba98b9e85d863672474ec9c81f